Add files via upload
Create a folder.  name of this folder To-Do App and inside this folder, create the mentioned files.
Create an index.html file. The file name must be index and its extension .html
Create a index.css file. The file name must be style and its extension .css
Create a index.js file. The file name must be script and its extension .js